// src/app/modules/OfferedCourse/OfferedCourse.interface.ts
import { Types } from "mongoose"

export type Days = "Sat" | "Sun" | "Mon" | "Tue" | "Wed" | "Thu" | "Fri"

export type TOfferedCourse = {
    semesterRegistration: Types.ObjectId;
    academicSemester: Types.ObjectId;
    academicDepartment: Types.ObjectId;
    course: Types.ObjectId;
    faculty: Types.ObjectId;
    maxCapacity: number;
    section: number;
    days: Days[];
    startTime: string;
    endTime: string;
}

export type TSchedule = {
    days: Days[] | undefined;
    startTime: string | undefined ;
    endTime: string | undefined;
}